+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.1.0
+- Fixed button height issue on resize.
+- Removed `countOffset` property, using css class `with-counter-desktop` instead for the button with counter height.
+
 ## 2.0.1
 - Added `popup` option for opening links in popup or not.